backend/requirements.txt
aniso8601==9.0.1
apimatic-core==0.2.3
apimatic-core-interfaces==0.1.3
apimatic-requests-client-adapter==0.1.4
blessed==1.20.0
blinker==1.6.2
CacheControl==0.12.14
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
deprecation==2.1.0
docutils==0.20.1
enum34==1.1.10
Faker==19.2.0
Flask==2.3.2
Flask-RESTful==0.3.10
idna==3.4
inquirer==3.1.3
itsdangerous==2.1.2
jsonpickle==3.0.1
jsonpointer==2.3
MarkupSafe==2.1.3
msgpack==1.0.5
packaging==23.1
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
pytz==2023.3
pywatchman==1.4.1
ratelimit==2.2.1
readchar==4.0.5
requests==2.31.0
six==1.16.0
squareup==33.0.0.20231115
urllib3==2.0.3
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.3.6
